Join two tables based on a distance metric of one or more columns

This differs from difference_join in that it considers all of the columns together when computing distance. This allows it to use metrics such as Euclidean or Manhattan that depend on multiple columns. Note that if you are computing with longitude or latitude, you probably want to use geo_join.

distance_join(
  x,
  y,
  by = NULL,
  max_dist = 1,
  method = c("euclidean", "manhattan"),
  mode = "inner",
  distance_col = NULL
)

distance_inner_join(
  x,
  y,
  by = NULL,
  method = "euclidean",
  max_dist = 1,
  distance_col = NULL
)

distance_left_join(
  x,
  y,
  by = NULL,
  method = "euclidean",
  max_dist = 1,
  distance_col = NULL
)

distance_right_join(
  x,
  y,
  by = NULL,
  method = "euclidean",
  max_dist = 1,
  distance_col = NULL
)

distance_full_join(
  x,
  y,
  by = NULL,
  method = "euclidean",
  max_dist = 1,
  distance_col = NULL
)

distance_semi_join(
  x,
  y,
  by = NULL,
  method = "euclidean",
  max_dist = 1,
  distance_col = NULL
)

distance_anti_join(
  x,
  y,
  by = NULL,
  method = "euclidean",
  max_dist = 1,
  distance_col = NULL
)

Arguments

x
A tbl

y
A tbl

by
Columns by which to join the two tables

max_dist
Maximum distance to use for joining

method
Method to use for computing distance, either euclidean (default) or manhattan.

mode
One of "inner", "left", "right", "full" "semi", or "anti"

distance_col
If given, will add a column with this name containing the distance between the two

Examples

library(dplyr)

head(iris)

#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa

sepal_lengths <- data_frame(Sepal.Length = c(5, 6, 7),
                            Sepal.Width = 1:3)

iris %>%
  distance_inner_join(sepal_lengths, max_dist = 2)

#> Joining by: c("Sepal.Length", "Sepal.Width")

#>     Sepal.Length.x Sepal.Width.x Petal.Length Petal.Width    Species
#> 1              5.1           3.5          1.4         0.2     setosa
#> 2              5.1           3.5          1.4         0.2     setosa
#> 3              4.9           3.0          1.4         0.2     setosa
#> 4              4.7           3.2          1.3         0.2     setosa
#> 5              4.6           3.1          1.5         0.2     setosa
#> 6              5.0           3.6          1.4         0.2     setosa
#> 7              5.4           3.9          1.7         0.4     setosa
#> 8              5.4           3.9          1.7         0.4     setosa
#> 9              4.6           3.4          1.4         0.3     setosa
#> 10             5.0           3.4          1.5         0.2     setosa
#> 11             4.4           2.9          1.4         0.2     setosa
#> 12             4.4           2.9          1.4         0.2     setosa
#> 13             4.9           3.1          1.5         0.1     setosa
#> 14             5.4           3.7          1.5         0.2     setosa
#> 15             5.4           3.7          1.5         0.2     setosa
#> 16             4.8           3.4          1.6         0.2     setosa
#> 17             4.8           3.0          1.4         0.1     setosa
#> 18             4.3           3.0          1.1         0.1     setosa
#> 19             5.8           4.0          1.2         0.2     setosa
#> 20             5.7           4.4          1.5         0.4     setosa
#> 21             5.4           3.9          1.3         0.4     setosa
#> 22             5.4           3.9          1.3         0.4     setosa
#> 23             5.1           3.5          1.4         0.3     setosa
#> 24             5.1           3.5          1.4         0.3     setosa
#> 25             5.7           3.8          1.7         0.3     setosa
#> 26             5.7           3.8          1.7         0.3     setosa
#> 27             5.4           3.4          1.7         0.2     setosa
#> 28             5.4           3.4          1.7         0.2     setosa
#> 29             5.1           3.7          1.5         0.4     setosa
#> 30             5.1           3.3          1.7         0.5     setosa
#> 31             5.1           3.3          1.7         0.5     setosa
#> 32             4.8           3.4          1.9         0.2     setosa
#> 33             5.0           3.0          1.6         0.2     setosa
#> 34             5.0           3.0          1.6         0.2     setosa
#> 35             5.0           3.0          1.6         0.2     setosa
#> 36             5.0           3.4          1.6         0.4     setosa
#> 37             5.2           3.5          1.5         0.2     setosa
#> 38             5.2           3.5          1.5         0.2     setosa
#> 39             5.2           3.4          1.4         0.2     setosa
#> 40             5.2           3.4          1.4         0.2     setosa
#> 41             4.7           3.2          1.6         0.2     setosa
#> 42             4.8           3.1          1.6         0.2     setosa
#> 43             5.4           3.4          1.5         0.4     setosa
#> 44             5.4           3.4          1.5         0.4     setosa
#> 45             5.5           4.2          1.4         0.2     setosa
#> 46             4.9           3.1          1.5         0.2     setosa
#> 47             5.0           3.2          1.2         0.2     setosa
#> 48             5.5           3.5          1.3         0.2     setosa
#> 49             5.5           3.5          1.3         0.2     setosa
#> 50             4.9           3.6          1.4         0.1     setosa
#> 51             4.4           3.0          1.3         0.2     setosa
#> 52             5.1           3.4          1.5         0.2     setosa
#> 53             5.1           3.4          1.5         0.2     setosa
#> 54             5.0           3.5          1.3         0.3     setosa
#> 55             4.5           2.3          1.3         0.3     setosa
#> 56             4.5           2.3          1.3         0.3     setosa
#> 57             4.4           3.2          1.3         0.2     setosa
#> 58             5.0           3.5          1.6         0.6     setosa
#> 59             4.8           3.0          1.4         0.3     setosa
#> 60             4.6           3.2          1.4         0.2     setosa
#> 61             5.3           3.7          1.5         0.2     setosa
#> 62             5.3           3.7          1.5         0.2     setosa
#> 63             5.0           3.3          1.4         0.2     setosa
#> 64             7.0           3.2          4.7         1.4 versicolor
#> 65             7.0           3.2          4.7         1.4 versicolor
#> 66             6.4           3.2          4.5         1.5 versicolor
#> 67             6.4           3.2          4.5         1.5 versicolor
#> 68             6.9           3.1          4.9         1.5 versicolor
#> 69             6.9           3.1          4.9         1.5 versicolor
#> 70             5.5           2.3          4.0         1.3 versicolor
#> 71             5.5           2.3          4.0         1.3 versicolor
#> 72             5.5           2.3          4.0         1.3 versicolor
#> 73             6.5           2.8          4.6         1.5 versicolor
#> 74             6.5           2.8          4.6         1.5 versicolor
#> 75             5.7           2.8          4.5         1.3 versicolor
#> 76             5.7           2.8          4.5         1.3 versicolor
#> 77             5.7           2.8          4.5         1.3 versicolor
#> 78             6.3           3.3          4.7         1.6 versicolor
#> 79             6.3           3.3          4.7         1.6 versicolor
#> 80             4.9           2.4          3.3         1.0 versicolor
#> 81             4.9           2.4          3.3         1.0 versicolor
#> 82             6.6           2.9          4.6         1.3 versicolor
#> 83             6.6           2.9          4.6         1.3 versicolor
#> 84             5.2           2.7          3.9         1.4 versicolor
#> 85             5.2           2.7          3.9         1.4 versicolor
#> 86             5.2           2.7          3.9         1.4 versicolor
#> 87             5.0           2.0          3.5         1.0 versicolor
#> 88             5.0           2.0          3.5         1.0 versicolor
#> 89             5.9           3.0          4.2         1.5 versicolor
#> 90             5.9           3.0          4.2         1.5 versicolor
#> 91             6.0           2.2          4.0         1.0 versicolor
#> 92             6.0           2.2          4.0         1.0 versicolor
#> 93             6.0           2.2          4.0         1.0 versicolor
#> 94             6.1           2.9          4.7         1.4 versicolor
#> 95             6.1           2.9          4.7         1.4 versicolor
#> 96             5.6           2.9          3.6         1.3 versicolor
#> 97             5.6           2.9          3.6         1.3 versicolor
#> 98             5.6           2.9          3.6         1.3 versicolor
#> 99             6.7           3.1          4.4         1.4 versicolor
#> 100            6.7           3.1          4.4         1.4 versicolor
#> 101            5.6           3.0          4.5         1.5 versicolor
#> 102            5.6           3.0          4.5         1.5 versicolor
#> 103            5.8           2.7          4.1         1.0 versicolor
#> 104            5.8           2.7          4.1         1.0 versicolor
#> 105            5.8           2.7          4.1         1.0 versicolor
#> 106            6.2           2.2          4.5         1.5 versicolor
#> 107            6.2           2.2          4.5         1.5 versicolor
#> 108            6.2           2.2          4.5         1.5 versicolor
#> 109            5.6           2.5          3.9         1.1 versicolor
#> 110            5.6           2.5          3.9         1.1 versicolor
#> 111            5.6           2.5          3.9         1.1 versicolor
#> 112            5.9           3.2          4.8         1.8 versicolor
#> 113            5.9           3.2          4.8         1.8 versicolor
#> 114            6.1           2.8          4.0         1.3 versicolor
#> 115            6.1           2.8          4.0         1.3 versicolor
#> 116            6.3           2.5          4.9         1.5 versicolor
#> 117            6.3           2.5          4.9         1.5 versicolor
#> 118            6.3           2.5          4.9         1.5 versicolor
#> 119            6.1           2.8          4.7         1.2 versicolor
#> 120            6.1           2.8          4.7         1.2 versicolor
#> 121            6.4           2.9          4.3         1.3 versicolor
#> 122            6.4           2.9          4.3         1.3 versicolor
#> 123            6.6           3.0          4.4         1.4 versicolor
#> 124            6.6           3.0          4.4         1.4 versicolor
#> 125            6.8           2.8          4.8         1.4 versicolor
#> 126            6.8           2.8          4.8         1.4 versicolor
#> 127            6.7           3.0          5.0         1.7 versicolor
#> 128            6.7           3.0          5.0         1.7 versicolor
#> 129            6.0           2.9          4.5         1.5 versicolor
#> 130            6.0           2.9          4.5         1.5 versicolor
#> 131            5.7           2.6          3.5         1.0 versicolor
#> 132            5.7           2.6          3.5         1.0 versicolor
#> 133            5.7           2.6          3.5         1.0 versicolor
#> 134            5.5           2.4          3.8         1.1 versicolor
#> 135            5.5           2.4          3.8         1.1 versicolor
#> 136            5.5           2.4          3.8         1.1 versicolor
#> 137            5.5           2.4          3.7         1.0 versicolor
#> 138            5.5           2.4          3.7         1.0 versicolor
#> 139            5.5           2.4          3.7         1.0 versicolor
#> 140            5.8           2.7          3.9         1.2 versicolor
#> 141            5.8           2.7          3.9         1.2 versicolor
#> 142            5.8           2.7          3.9         1.2 versicolor
#> 143            6.0           2.7          5.1         1.6 versicolor
#> 144            6.0           2.7          5.1         1.6 versicolor
#> 145            6.0           2.7          5.1         1.6 versicolor
#> 146            5.4           3.0          4.5         1.5 versicolor
#> 147            5.4           3.0          4.5         1.5 versicolor
#> 148            6.0           3.4          4.5         1.6 versicolor
#> 149            6.0           3.4          4.5         1.6 versicolor
#> 150            6.7           3.1          4.7         1.5 versicolor
#> 151            6.7           3.1          4.7         1.5 versicolor
#> 152            6.3           2.3          4.4         1.3 versicolor
#> 153            6.3           2.3          4.4         1.3 versicolor
#> 154            6.3           2.3          4.4         1.3 versicolor
#> 155            5.6           3.0          4.1         1.3 versicolor
#> 156            5.6           3.0          4.1         1.3 versicolor
#> 157            5.5           2.5          4.0         1.3 versicolor
#> 158            5.5           2.5          4.0         1.3 versicolor
#> 159            5.5           2.5          4.0         1.3 versicolor
#> 160            5.5           2.6          4.4         1.2 versicolor
#> 161            5.5           2.6          4.4         1.2 versicolor
#> 162            5.5           2.6          4.4         1.2 versicolor
#> 163            6.1           3.0          4.6         1.4 versicolor
#> 164            6.1           3.0          4.6         1.4 versicolor
#> 165            5.8           2.6          4.0         1.2 versicolor
#> 166            5.8           2.6          4.0         1.2 versicolor
#> 167            5.8           2.6          4.0         1.2 versicolor
#> 168            5.0           2.3          3.3         1.0 versicolor
#> 169            5.0           2.3          3.3         1.0 versicolor
#> 170            5.6           2.7          4.2         1.3 versicolor
#> 171            5.6           2.7          4.2         1.3 versicolor
#> 172            5.6           2.7          4.2         1.3 versicolor
#> 173            5.7           3.0          4.2         1.2 versicolor
#> 174            5.7           3.0          4.2         1.2 versicolor
#> 175            5.7           2.9          4.2         1.3 versicolor
#> 176            5.7           2.9          4.2         1.3 versicolor
#> 177            6.2           2.9          4.3         1.3 versicolor
#> 178            6.2           2.9          4.3         1.3 versicolor
#> 179            5.1           2.5          3.0         1.1 versicolor
#> 180            5.1           2.5          3.0         1.1 versicolor
#> 181            5.1           2.5          3.0         1.1 versicolor
#> 182            5.7           2.8          4.1         1.3 versicolor
#> 183            5.7           2.8          4.1         1.3 versicolor
#> 184            5.7           2.8          4.1         1.3 versicolor
#> 185            6.3           3.3          6.0         2.5  virginica
#> 186            6.3           3.3          6.0         2.5  virginica
#> 187            5.8           2.7          5.1         1.9  virginica
#> 188            5.8           2.7          5.1         1.9  virginica
#> 189            5.8           2.7          5.1         1.9  virginica
#> 190            7.1           3.0          5.9         2.1  virginica
#> 191            7.1           3.0          5.9         2.1  virginica
#> 192            6.3           2.9          5.6         1.8  virginica
#> 193            6.3           2.9          5.6         1.8  virginica
#> 194            6.5           3.0          5.8         2.2  virginica
#> 195            6.5           3.0          5.8         2.2  virginica
#> 196            7.6           3.0          6.6         2.1  virginica
#> 197            7.6           3.0          6.6         2.1  virginica
#> 198            4.9           2.5          4.5         1.7  virginica
#> 199            4.9           2.5          4.5         1.7  virginica
#> 200            7.3           2.9          6.3         1.8  virginica
#> 201            7.3           2.9          6.3         1.8  virginica
#> 202            6.7           2.5          5.8         1.8  virginica
#> 203            6.7           2.5          5.8         1.8  virginica
#> 204            7.2           3.6          6.1         2.5  virginica
#> 205            7.2           3.6          6.1         2.5  virginica
#> 206            6.5           3.2          5.1         2.0  virginica
#> 207            6.5           3.2          5.1         2.0  virginica
#> 208            6.4           2.7          5.3         1.9  virginica
#> 209            6.4           2.7          5.3         1.9  virginica
#> 210            6.8           3.0          5.5         2.1  virginica
#> 211            6.8           3.0          5.5         2.1  virginica
#> 212            5.7           2.5          5.0         2.0  virginica
#> 213            5.7           2.5          5.0         2.0  virginica
#> 214            5.7           2.5          5.0         2.0  virginica
#> 215            5.8           2.8          5.1         2.4  virginica
#> 216            5.8           2.8          5.1         2.4  virginica
#> 217            5.8           2.8          5.1         2.4  virginica
#> 218            6.4           3.2          5.3         2.3  virginica
#> 219            6.4           3.2          5.3         2.3  virginica
#> 220            6.5           3.0          5.5         1.8  virginica
#> 221            6.5           3.0          5.5         1.8  virginica
#> 222            7.7           3.8          6.7         2.2  virginica
#> 223            7.7           2.6          6.9         2.3  virginica
#> 224            7.7           2.6          6.9         2.3  virginica
#> 225            6.0           2.2          5.0         1.5  virginica
#> 226            6.0           2.2          5.0         1.5  virginica
#> 227            6.0           2.2          5.0         1.5  virginica
#> 228            6.9           3.2          5.7         2.3  virginica
#> 229            6.9           3.2          5.7         2.3  virginica
#> 230            5.6           2.8          4.9         2.0  virginica
#> 231            5.6           2.8          4.9         2.0  virginica
#> 232            5.6           2.8          4.9         2.0  virginica
#> 233            7.7           2.8          6.7         2.0  virginica
#> 234            7.7           2.8          6.7         2.0  virginica
#> 235            6.3           2.7          4.9         1.8  virginica
#> 236            6.3           2.7          4.9         1.8  virginica
#> 237            6.7           3.3          5.7         2.1  virginica
#> 238            6.7           3.3          5.7         2.1  virginica
#> 239            7.2           3.2          6.0         1.8  virginica
#> 240            7.2           3.2          6.0         1.8  virginica
#> 241            6.2           2.8          4.8         1.8  virginica
#> 242            6.2           2.8          4.8         1.8  virginica
#> 243            6.1           3.0          4.9         1.8  virginica
#> 244            6.1           3.0          4.9         1.8  virginica
#> 245            6.4           2.8          5.6         2.1  virginica
#> 246            6.4           2.8          5.6         2.1  virginica
#> 247            7.2           3.0          5.8         1.6  virginica
#> 248            7.2           3.0          5.8         1.6  virginica
#> 249            7.4           2.8          6.1         1.9  virginica
#> 250            7.4           2.8          6.1         1.9  virginica
#> 251            7.9           3.8          6.4         2.0  virginica
#> 252            6.4           2.8          5.6         2.2  virginica
#> 253            6.4           2.8          5.6         2.2  virginica
#> 254            6.3           2.8          5.1         1.5  virginica
#> 255            6.3           2.8          5.1         1.5  virginica
#> 256            6.1           2.6          5.6         1.4  virginica
#> 257            6.1           2.6          5.6         1.4  virginica
#> 258            6.1           2.6          5.6         1.4  virginica
#> 259            7.7           3.0          6.1         2.3  virginica
#> 260            7.7           3.0          6.1         2.3  virginica
#> 261            6.3           3.4          5.6         2.4  virginica
#> 262            6.3           3.4          5.6         2.4  virginica
#> 263            6.4           3.1          5.5         1.8  virginica
#> 264            6.4           3.1          5.5         1.8  virginica
#> 265            6.0           3.0          4.8         1.8  virginica
#> 266            6.0           3.0          4.8         1.8  virginica
#> 267            6.9           3.1          5.4         2.1  virginica
#> 268            6.9           3.1          5.4         2.1  virginica
#> 269            6.7           3.1          5.6         2.4  virginica
#> 270            6.7           3.1          5.6         2.4  virginica
#> 271            6.9           3.1          5.1         2.3  virginica
#> 272            6.9           3.1          5.1         2.3  virginica
#> 273            5.8           2.7          5.1         1.9  virginica
#> 274            5.8           2.7          5.1         1.9  virginica
#> 275            5.8           2.7          5.1         1.9  virginica
#> 276            6.8           3.2          5.9         2.3  virginica
#> 277            6.8           3.2          5.9         2.3  virginica
#> 278            6.7           3.3          5.7         2.5  virginica
#> 279            6.7           3.3          5.7         2.5  virginica
#> 280            6.7           3.0          5.2         2.3  virginica
#> 281            6.7           3.0          5.2         2.3  virginica
#> 282            6.3           2.5          5.0         1.9  virginica
#> 283            6.3           2.5          5.0         1.9  virginica
#> 284            6.3           2.5          5.0         1.9  virginica
#> 285            6.5           3.0          5.2         2.0  virginica
#> 286            6.5           3.0          5.2         2.0  virginica
#> 287            6.2           3.4          5.4         2.3  virginica
#> 288            6.2           3.4          5.4         2.3  virginica
#> 289            5.9           3.0          5.1         1.8  virginica
#> 290            5.9           3.0          5.1         1.8  virginica
#>     Sepal.Length.y Sepal.Width.y
#> 1                6             2
#> 2                7             3
#> 3                6             2
#> 4                6             2
#> 5                6             2
#> 6                6             2
#> 7                6             2
#> 8                7             3
#> 9                6             2
#> 10               6             2
#> 11               5             1
#> 12               6             2
#> 13               6             2
#> 14               6             2
#> 15               7             3
#> 16               6             2
#> 17               6             2
#> 18               6             2
#> 19               7             3
#> 20               7             3
#> 21               6             2
#> 22               7             3
#> 23               6             2
#> 24               7             3
#> 25               6             2
#> 26               7             3
#> 27               6             2
#> 28               7             3
#> 29               6             2
#> 30               6             2
#> 31               7             3
#> 32               6             2
#> 33               5             1
#> 34               6             2
#> 35               7             3
#> 36               6             2
#> 37               6             2
#> 38               7             3
#> 39               6             2
#> 40               7             3
#> 41               6             2
#> 42               6             2
#> 43               6             2
#> 44               7             3
#> 45               7             3
#> 46               6             2
#> 47               6             2
#> 48               6             2
#> 49               7             3
#> 50               6             2
#> 51               6             2
#> 52               6             2
#> 53               7             3
#> 54               6             2
#> 55               5             1
#> 56               6             2
#> 57               6             2
#> 58               6             2
#> 59               6             2
#> 60               6             2
#> 61               6             2
#> 62               7             3
#> 63               6             2
#> 64               6             2
#> 65               7             3
#> 66               6             2
#> 67               7             3
#> 68               6             2
#> 69               7             3
#> 70               5             1
#> 71               6             2
#> 72               7             3
#> 73               6             2
#> 74               7             3
#> 75               5             1
#> 76               6             2
#> 77               7             3
#> 78               6             2
#> 79               7             3
#> 80               5             1
#> 81               6             2
#> 82               6             2
#> 83               7             3
#> 84               5             1
#> 85               6             2
#> 86               7             3
#> 87               5             1
#> 88               6             2
#> 89               6             2
#> 90               7             3
#> 91               5             1
#> 92               6             2
#> 93               7             3
#> 94               6             2
#> 95               7             3
#> 96               5             1
#> 97               6             2
#> 98               7             3
#> 99               6             2
#> 100              7             3
#> 101              6             2
#> 102              7             3
#> 103              5             1
#> 104              6             2
#> 105              7             3
#> 106              5             1
#> 107              6             2
#> 108              7             3
#> 109              5             1
#> 110              6             2
#> 111              7             3
#> 112              6             2
#> 113              7             3
#> 114              6             2
#> 115              7             3
#> 116              5             1
#> 117              6             2
#> 118              7             3
#> 119              6             2
#> 120              7             3
#> 121              6             2
#> 122              7             3
#> 123              6             2
#> 124              7             3
#> 125              6             2
#> 126              7             3
#> 127              6             2
#> 128              7             3
#> 129              6             2
#> 130              7             3
#> 131              5             1
#> 132              6             2
#> 133              7             3
#> 134              5             1
#> 135              6             2
#> 136              7             3
#> 137              5             1
#> 138              6             2
#> 139              7             3
#> 140              5             1
#> 141              6             2
#> 142              7             3
#> 143              5             1
#> 144              6             2
#> 145              7             3
#> 146              6             2
#> 147              7             3
#> 148              6             2
#> 149              7             3
#> 150              6             2
#> 151              7             3
#> 152              5             1
#> 153              6             2
#> 154              7             3
#> 155              6             2
#> 156              7             3
#> 157              5             1
#> 158              6             2
#> 159              7             3
#> 160              5             1
#> 161              6             2
#> 162              7             3
#> 163              6             2
#> 164              7             3
#> 165              5             1
#> 166              6             2
#> 167              7             3
#> 168              5             1
#> 169              6             2
#> 170              5             1
#> 171              6             2
#> 172              7             3
#> 173              6             2
#> 174              7             3
#> 175              6             2
#> 176              7             3
#> 177              6             2
#> 178              7             3
#> 179              5             1
#> 180              6             2
#> 181              7             3
#> 182              5             1
#> 183              6             2
#> 184              7             3
#> 185              6             2
#> 186              7             3
#> 187              5             1
#> 188              6             2
#> 189              7             3
#> 190              6             2
#> 191              7             3
#> 192              6             2
#> 193              7             3
#> 194              6             2
#> 195              7             3
#> 196              6             2
#> 197              7             3
#> 198              5             1
#> 199              6             2
#> 200              6             2
#> 201              7             3
#> 202              6             2
#> 203              7             3
#> 204              6             2
#> 205              7             3
#> 206              6             2
#> 207              7             3
#> 208              6             2
#> 209              7             3
#> 210              6             2
#> 211              7             3
#> 212              5             1
#> 213              6             2
#> 214              7             3
#> 215              5             1
#> 216              6             2
#> 217              7             3
#> 218              6             2
#> 219              7             3
#> 220              6             2
#> 221              7             3
#> 222              7             3
#> 223              6             2
#> 224              7             3
#> 225              5             1
#> 226              6             2
#> 227              7             3
#> 228              6             2
#> 229              7             3
#> 230              5             1
#> 231              6             2
#> 232              7             3
#> 233              6             2
#> 234              7             3
#> 235              6             2
#> 236              7             3
#> 237              6             2
#> 238              7             3
#> 239              6             2
#> 240              7             3
#> 241              6             2
#> 242              7             3
#> 243              6             2
#> 244              7             3
#> 245              6             2
#> 246              7             3
#> 247              6             2
#> 248              7             3
#> 249              6             2
#> 250              7             3
#> 251              7             3
#> 252              6             2
#> 253              7             3
#> 254              6             2
#> 255              7             3
#> 256              5             1
#> 257              6             2
#> 258              7             3
#> 259              6             2
#> 260              7             3
#> 261              6             2
#> 262              7             3
#> 263              6             2
#> 264              7             3
#> 265              6             2
#> 266              7             3
#> 267              6             2
#> 268              7             3
#> 269              6             2
#> 270              7             3
#> 271              6             2
#> 272              7             3
#> 273              5             1
#> 274              6             2
#> 275              7             3
#> 276              6             2
#> 277              7             3
#> 278              6             2
#> 279              7             3
#> 280              6             2
#> 281              7             3
#> 282              5             1
#> 283              6             2
#> 284              7             3
#> 285              6             2
#> 286              7             3
#> 287              6             2
#> 288              7             3
#> 289              6             2
#> 290              7             3