Join two tables based on a distance metric of one or more columns

This differs from difference_join in that it considers all of the columns together when computing distance. This allows it to use metrics such as Euclidean or Manhattan that depend on multiple columns. Note that if you are computing with longitude or latitude, you probably want to use geo_join.

distance_join(x, y, by = NULL, max_dist = 1, method = c("euclidean",
  "manhattan"), mode = "inner", distance_col = NULL)

distance_inner_join(x, y, by = NULL, method = "euclidean", max_dist = 1,
  distance_col = NULL)

distance_left_join(x, y, by = NULL, method = "euclidean", max_dist = 1,
  distance_col = NULL)

distance_right_join(x, y, by = NULL, method = "euclidean", max_dist = 1,
  distance_col = NULL)

distance_full_join(x, y, by = NULL, method = "euclidean", max_dist = 1,
  distance_col = NULL)

distance_semi_join(x, y, by = NULL, method = "euclidean", max_dist = 1,
  distance_col = NULL)

distance_anti_join(x, y, by = NULL, method = "euclidean", max_dist = 1,
  distance_col = NULL)

Arguments

x
A tbl

y
A tbl

by
Columns by which to join the two tables

max_dist
Maximum distance to use for joining

method
Method to use for computing distance, either euclidean (default) or manhattan.

mode
One of "inner", "left", "right", "full" "semi", or "anti"

distance_col
If given, will add a column with this name containing the distance between the two

Examples

library(dplyr)

head(iris)

#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa

sepal_lengths <- data_frame(Sepal.Length = c(5, 6, 7),
                            Sepal.Width = 1:3)

iris %>%
  distance_inner_join(sepal_lengths, max_dist = 2)

#> Joining by: c("Sepal.Length", "Sepal.Width")

#>     Sepal.Length.x Sepal.Width.x Petal.Length Petal.Width    Species
#> 1              5.1           3.5          1.4         0.2     setosa
#> 2              5.1           3.5          1.4         0.2     setosa
#> 3              4.9           3.0          1.4         0.2     setosa
#> 4              4.7           3.2          1.3         0.2     setosa
#> 5              4.6           3.1          1.5         0.2     setosa
#> 6              5.0           3.6          1.4         0.2     setosa
#> 7              5.4           3.9          1.7         0.4     setosa
#> 8              5.4           3.9          1.7         0.4     setosa
#> 9              4.6           3.4          1.4         0.3     setosa
#> 10             5.0           3.4          1.5         0.2     setosa
#> 11             4.4           2.9          1.4         0.2     setosa
#> 12             4.4           2.9          1.4         0.2     setosa
#> 13             4.9           3.1          1.5         0.1     setosa
#> 14             5.4           3.7          1.5         0.2     setosa
#> 15             5.4           3.7          1.5         0.2     setosa
#> 16             4.8           3.4          1.6         0.2     setosa
#> 17             4.8           3.0          1.4         0.1     setosa
#> 18             4.3           3.0          1.1         0.1     setosa
#> 19             5.8           4.0          1.2         0.2     setosa
#> 20             5.7           4.4          1.5         0.4     setosa
#> 21             5.4           3.9          1.3         0.4     setosa
#> 22             5.4           3.9          1.3         0.4     setosa
#> 23             5.1           3.5          1.4         0.3     setosa
#> 24             5.1           3.5          1.4         0.3     setosa
#> 25             5.7           3.8          1.7         0.3     setosa
#> 26             5.7           3.8          1.7         0.3     setosa
#> 27             5.4           3.4          1.7         0.2     setosa
#> 28             5.4           3.4          1.7         0.2     setosa
#>     Sepal.Length.y Sepal.Width.y
#> 1                6             2
#> 2                7             3
#> 3                6             2
#> 4                6             2
#> 5                6             2
#> 6                6             2
#> 7                6             2
#> 8                7             3
#> 9                6             2
#> 10               6             2
#> 11               5             1
#> 12               6             2
#> 13               6             2
#> 14               6             2
#> 15               7             3
#> 16               6             2
#> 17               6             2
#> 18               6             2
#> 19               7             3
#> 20               7             3
#> 21               6             2
#> 22               7             3
#> 23               6             2
#> 24               7             3
#> 25               6             2
#> 26               7             3
#> 27               6             2
#> 28               7             3
#>  [ reached getOption("max.print") -- omitted 262 rows ]